requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
beautifulsoup4
praw
python-dotenv
openai
google-cloud-texttospeech
duckduckgo-search
pypdf
python-docx

celery
redis
google-auth
